User can now apply to open job offers, minor bugs fixed!
The user can now apply to job offers, if they are sign in or have an account. updated the education end date so the user see 'Currently study here' if he had set it to currently studying here. User's job applications are fetched but still have to work on displaying them.